finish the program

+++ src/main/java/org/zipcode/luckyMoney/Repository/LuckymoneyRepository.java
@@ -3,11 +3,12 @@ package org.zipcode.luckyMoney.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.zipcode.luckyMoney.Entity.Luckymoney;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface LuckymoneyRepository extends JpaRepository<Luckymoney,Integer> {
 
 
 
-    public List<Luckymoney> findByMoney(double d);
+    public List<Luckymoney> findByMoney(BigDecimal d);
 }
